# api/public/swagger.yaml
components:
...
    schemas:
...
        UserResponse:
            properties:
                id:
                    type: number
                    format: double
                username:
                    type: string
                displayName:
                    type: string
            required:
                - id
                - username
                - displayName
            type: object
            additionalProperties: false
        GroupResponse:
            properties:
                groupId:
                    type: number
                    format: double
                name:
                    type: string
                description:
                    type: string
            required:
                - groupId
                - name
                - description
            type: object
            additionalProperties: false
        UserDetailsResponse:
            properties:
                id:
                    type: number
                    format: double
                displayName:
                    type: string
                username:
                    type: string
                isActive:
                    type: boolean
                isAdmin:
                    type: boolean
                autoExec:
                    type: string
                groups:
                    items:
                        $ref: '#/components/schemas/GroupResponse'
                    type: array
            required:
                - id
                - displayName
                - username
                - isActive
                - isAdmin
            type: object
            additionalProperties: false
        UserPayload:
            properties:
                displayName:
                    type: string
                    description: 'Display name for user'
                    example: 'John Snow'
                username:
                    type: string
                    description: 'Username for user'
                    example: johnSnow01
                password:
                    type: string
                    description: 'Password for user'
                isAdmin:
                    type: boolean
                    description: 'Account should be admin or not, defaults to false'
                    example: 'false'
                isActive:
                    type: boolean
                    description: 'Account should be active or not, defaults to true'
                    example: 'true'
                autoExec:
                    type: string
                    description: 'User-specific auto-exec code'
                    example: ""
            required:
                - displayName
                - username
                - password
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: api
    version: 0.0.2
    description: 'Api of SASjs server'
    contact:
        name: '4GL Ltd'
openapi: 3.0.0
paths:
...
    /SASjsApi/user:
        get:
            operationId: GetAllUsers
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/UserResponse'
                                type: array
                            examples:
                                'Example 1':
                                    value: [{id: 123, username: johnusername, displayName: John}, {id: 456, username: starkusername, displayName: Stark}]
            summary: 'Get list of all users (username, displayname). All users can request this.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters: []
        post:
            operationId: CreateUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailsResponse'
                            examples:
                                'Example 1':
                                    value: {id: 1234, displayName: 'John Snow', username: johnSnow01, isAdmin: false, isActive: true}
            summary: 'Create user with the following attributes: UserId, UserName, Password, isAdmin, isActive. Admin only task.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserPayload'
    '/SASjsApi/user/by/username/{username}':
        get:
            operationId: GetUserByUsername
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailsResponse'
            description: 'Only Admin or user itself will get user autoExec code.'
            summary: 'Get user properties - such as group memberships, userName, displayName.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'The User''s username'
                    in: path
                    name: username
                    required: true
                    schema:
                        type: string
                    example: johnSnow01
        patch:
            operationId: UpdateUserByUsername
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailsResponse'
                            examples:
                                'Example 1':
                                    value: {id: 1234, displayName: 'John Snow', username: johnSnow01, isAdmin: false, isActive: true}
            summary: 'Update user properties - such as displayName. Can be performed either by admins, or the user in question.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'The User''s username'
                    in: path
                    name: username
                    required: true
                    schema:
                        type: string
                    example: johnSnow01
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserPayload'
        delete:
            operationId: DeleteUserByUsername
            responses:
                '204':
                    description: 'No content'
            summary: 'Delete a user. Can be performed either by admins, or the user in question.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'The User''s username'
                    in: path
                    name: username
                    required: true
                    schema:
                        type: string
                    example: johnSnow01
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                            type: object
    '/SASjsApi/user/{userId}':
        get:
            operationId: GetUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailsResponse'
            description: 'Only Admin or user itself will get user autoExec code.'
            summary: 'Get user properties - such as group memberships, userName, displayName.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'The user''s identifier'
                    in: path
                    name: userId
                    required: true
                    schema:
                        format: double
                        type: number
                    example: 1234
        patch:
            operationId: UpdateUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailsResponse'
                            examples:
                                'Example 1':
                                    value: {id: 1234, displayName: 'John Snow', username: johnSnow01, isAdmin: false, isActive: true}
            summary: 'Update user properties - such as displayName. Can be performed either by admins, or the user in question.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'The user''s identifier'
                    in: path
                    name: userId
                    required: true
                    schema:
                        format: double
                        type: number
                    example: '1234'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserPayload'
        delete:
            operationId: DeleteUser
            responses:
                '204':
                    description: 'No content'
            summary: 'Delete a user. Can be performed either by admins, or the user in question.'
            tags:
                - User
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'The user''s identifier'
                    in: path
                    name: userId
                    required: true
                    schema:
                        format: double
                        type: number
                    example: 1234
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                            type: object
...
servers:
    -
        url: /
...